test(gui-app): share one hardened Chrome launcher across the browser drivers
#1552 hardened the Chrome launch in diff-edit-browser-regression.mjs after a
cold CI runner took Chrome past the flat 30s DevTools deadline and left a pile
of crashpad orphans to reap: three attempts, a fresh mkdtemp profile per
attempt, detached process-group ownership, a bounded tree-kill that VERIFIES
the group is gone, and a spawn-error listener so an ENOENT/EACCES surfaces as
a message instead of an uncaught event.

Its siblings kept the pre-#1552 copy - one attempt behind one flat deadline, a
debugging port picked in advance, kill("SIGKILL") aimed at the browser PID
alone, and a 300ms sleep standing in for "the profile is safe to delete now".
They run on the same runners in the same gate, so they flake the same way.

Extract the launcher to scripts/chrome-launcher.mjs and put every driver on
it. The module owns findChrome, launchChromeWithDevTools and
terminateProcessTree; each driver keeps its own fixture, CDP client and
assertions. Two launch details travel with it and are the point rather than
incidental: the debugging port is now 0 and the endpoint is read back off
Chrome's stderr, because a port picked in advance is only free until Chrome
binds it and on a retry is still held by the attempt that just died; and
teardown waits for the process GROUP to be verifiably gone, which is what the
sleep was approximating.

Two behaviours are unified rather than preserved per-driver, both widening:
CHROME_BIN is honoured everywhere (only diff-edit read it), and the candidate
list is the union, so the siblings keep Chromium.app and gain Chrome for
Testing.

It sits beside the drivers rather than under scripts/lib/ because .gitignore
ignores **/lib/ for build output - a launcher there is untracked, and CI fails
on a module the four drivers import and nobody pushed.
